Fix "Sussman and Jr." reference.
In the Preface, the reference to the AIM 349 Scheme paper rendered as
"Sussman and Jr."  It was taking "Jr." as the last name of "Guy
L. Steele Jr."  With this change (I added a comma), it now renders as
"Sussman and Guy L. Steele," which is better.  The authors may have a
better fix.