First argument to a function call should be placed in %rdi
There was a typo in the example shown in the assign homes pass. The generated assembly should put the argument for print_int into %rdi instead of %rax, before the callq instruction.

The assembly shown for the Python version in the patch instructions pass looks correct, but the racket version still had the typo.